requirements.txt
alembic==1.4.2
amqp==5.0.1
atomicwrites==1.4.0
attrs==19.3.0
Authlib==0.14.3
awscli==1.18.82
base58==2.0.1
billiard==3.6.3.0
bitarray==1.2.2
boto3==1.14.6
botocore==1.17.6
celery==5.0.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.3
cryptography==2.9.2
cytoolz==0.11.0
docutils==0.15.2
ecdsa==0.15
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.9.5
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
hexbytes==0.2.1
idna==2.9
importlib-metadata==1.6.0
ipfshttpclient==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
kombu==5.0.2
lru-dict==1.1.6
Mako==1.1.2
MarkupSafe==1.1.1
more-itertools==8.2.0
multiaddr==0.0.9
netaddr==0.8.0
packaging==20.3
parsimonious==0.8.1
pluggy==0.13.1
prompt-toolkit==3.0.8
protobuf==3.13.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
py==1.8.1
pyasn1==0.4.8
pycountry==19.8.18
pycparser==2.20
pycryptodome==3.9.8
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.17.3
pytest==5.4.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-jose==3.1.0
pytz==2020.1
pywin32==228
PyYAML==5.3.1
redis==3.5.3
requests==2.23.0
rlp==2.0.0
rsa==3.4.2
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.16
toolz==0.11.1
typing-extensions==3.7.4.3
urllib3==1.25.9
varint==1.0.2
vine==5.0.0
wcwidth==0.1.9
web3==5.12.2
websockets==8.1
Werkzeug==1.0.1
wincertstore==0.2
WTForms==2.3.1
zipp==3.1.0
